wip: refactor ProjectSettingItemsSeeder
