+++ src/ProjectSettingServiceProvider.php
@@ -26,7 +26,7 @@ class ProjectSettingServiceProvider extends ServiceProvider
 
     private $packageSeeders = [
         'ProjectSettingGroupsTableSeeder',
-        'ProjectSettingSectionsWithSettingItemsSeeder',
+        'ProjectSettingItemsTableSeeder',
     ];
 
     /**
